backend/requirements.txt
fastapi
uvicorn
python-multipart
gTTS
edge-tts
pyttsx3
groq
requests
websocket-client
gradio
google-auth-oauthlib
google-api-python-client
google-auth
